backend/schemas/users.py
from pydantic import BaseModel, EmailStr
import uuid


class UserCreate(BaseModel):
    email: EmailStr
    password: str


class UserRead(BaseModel):
    id: uuid.UUID
    email: EmailStr

    class Config:
        from_attributes = True  # Useful when returning sql models
