backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=3.0.2",
    "fastapi-mail>=1.5.0",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.8",
    "sqlmodel>=0.0.24",
]
